+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2025-01-22
+
+- Makes `Value` and `Row` JSON serializable.
+
 ## [0.2.1] - 2023-02-09
 
-- CsvParser casts to float on all values.
+- `CsvParser` casts to float on all values.
 
 ## [0.2.0] - 2023-02-01